+++ telnet_client.cpp
@@ -26,9 +26,9 @@ telnet_client::telnet_client(boost::asio::io_service& io_service,
 }
  
 // pass the write data to the do_write function via the io service in the other thread
-void telnet_client::write(const char* msg, int len)
+void telnet_client::write(const char* msg, size_t len)
 {
-  for(int i=0; i<len; i++)
+  for(unsigned i=0; i<len; i++)
   {
     io_service_.post(boost::bind(&telnet_client::do_write, this, msg[i]));
   }
